Added cascading lockouts Added functionality to allow for a component and all sub components to be disabled as well as enabled.
Currently this is used when automatically configuring ports.  When the
auto-configure button is pressed, everything in the application is
disbaled until the operation is complete.  This not only protects the
process from being disturbed by the user, but it also indicates to the
user that something actually happened, and when it has finished